Revert "change ids from integer to long"
This reverts commit 3b0f6ac21a40ada760f328568476173be2ba4fbb.

+++ backend/src/main/java/com/library/catalog/models/Book.java
@@ -2,11 +2,13 @@ package com.library.catalog.models;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.hibernate.Hibernate;
 
 import java.util.List;
 
@@ -22,7 +24,7 @@ public class Book {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", unique = true)
-    private Long id;
+    private Integer id;
 
     @Column(name = "title", unique = true, nullable = false)
     @NotBlank
